+++ src/main/java/com/CuteNekoDragon/Core/common/datagen/recipes/removeRecipes/RecipeRemovals.java
@@ -1,0 +1,10 @@
+package com.CuteNekoDragon.Core.common.datagen.recipes.removeRecipes;
+
+import com.CuteNekoDragon.Core.utils.RecipeBlacklist;
+
+public class RecipeRemovals {
+
+    public static void register() {
+        RecipeBlacklist.blacklistId("minecraft:golden_sword");
+    }
+}